Use abstract spec for dependencies in setup.py
`pip.req.parse_requirements` is not available in pip 10, so this package breaks when attempting to install it with pip 10. This change also makes the logical distinction between the abstract dependencies in setup.py and the concrete dependencies in requirements.txt.

I believe these to be reasonable version ranges based on what we have in cfgov-refresh already, what's specified in requirements.txt, and the changelog for each of the dependencies.