Fix intermittent test failures
Replace buildDefaultAggExclude with a(n ordered) list of agg_exclude
strings.

The buildDefaultAggExclude method was unnecessary. It allowed adding
additional strings to the agg_exclude collection for a search, but this
was not used in any of the specs in this file.

It was causing intermittent spec failures because it used a set (an
unordered collection) for the strings, but the spec assertions were
expecting a list in a specific order. Sometimes the agg_exclude list
would come in the wrong order and cause tests to fail.