Require library versions up-to-date with cfgov-refresh
urllib3 is not directly required in this repo, so allow elasticsearch to
pull in whatever version it wants.

In general, we allow library versions to be flexible within major versions,
for libraries that use semantic versioning. Some exceptions include
djangorestframework (since it drops django1.8 support after 3.6.4) and
requests (since our use of it in this repo is custom and more likely to
be broken by changes)