update version of django-flags to 4.0.x
As of cfgov-refresh#4641 [0], cf.gov uses version 4.0.1 of the
django-flags package. This package currently requires >=3.0.2,<4, and so
this requirement needs to be updated to remain compatible.

There are several breaking changes in django-flags 4.0 which need to be
addressed here as well:

1. Because of the way the django-flags 4.0 loads flag conditions from
the database, the "flags" app must be included in Django settings
INSTALLED_APPS. This commit modifies the test settings to do so.
2. If a flag is undefined, the flag_enabled() method now returns None
instead of False. This package has some tests and logic that assume that
the value is False, so this commit converts the return value to a
boolean.

All unit tests pass properly with this change.

[0] https://github.com/cfpb/cfgov-refresh/pull/4641