Run the linter as part of Travis testing
This change adds the linter commands (flake8/isort) to the Travis CI
configuration for this project.

To get the existing linter tests to pass, it sets lines_after_imports=2,
which feels like a reasonable setting that we already have in
cfgov-refresh.

This also adds .python-version to the .gitignore so that local
configuration of pyenv doesn't get tracked by git.