Properly support Unicode in CSV exports
Results containing non-ASCII characters will currently cause encoding
errors when a CSV export is triggered, which causes results to be cut
off prematurely.

This is due to the change in ccdb5-api#93 that removed use of a default
UTF-8 encoding.

This fix adds use of the unicodecsv Python package (already in use in
cfgov-refresh) so that exports work properly when using Python 2.

A new test has been added to verify this behavior.