Pin Django Rest Framework to <4.0
This change will pin Django Rest Framework to >=3.9.1 to detail with a security issue that GitHub has flagged, and <4.0 which is the the exclusive maximum that Wagtail 2.8 currently pins (which we have to be compatible with when we get to cfgov-refresh).

There's some minor incompatibility between the previous version and more recents in terms of the way error dicts are returned. The error strings are now wrapped in `ErrorDetail`. This change also updates the tests as appropriate.