Remove scroll search for ES7 and add search_after pagination
Elasticsearch no longer supports using a "from" value to paginate results of a
"scroll" search, so we need a new pagination method.

This implements Elasticsearch's recommended use of "search_after" for pagination.

Testing
This can be tested locally by following the INSTALL directions in the ccdb-data-pipeline repo and
getting a local index populated, or by deploying a version of cfgov to a dev server that
installs the es7-remove-scroll branch of ccdb5-api.