Sidestep API throttling when in DEBUG mode
Integration testing with Cypress is difficult when testing on a server, because many of the
test actions generate API calls that violate throttling rules, and the tests have to be
slowed down to work.

This change allows throttling to be sidestepped when DEBUG is True, which can be set
locally and on servers where CI testing could be run.