Fix CI tagging to have PEP 440-compliant version
On modern versions of pip, CI testing of this package fails because
it tries to build testing wheels with a version like

ci.dev0+61f3f55

As of PEP 440 (https://peps.python.org/pep-0440/) though, this is no
longer a valid public version identifier, as the first element must be
a number.

This commit alters the bespoke versioning logic in setup.py to instead
use the number 0, e.g.

0.dev0+61f3f55

Best I can tell, this version number is only used when CI testing
on GitHub Actions against a commit SHA, for example:

https://github.com/cfpb/ccdb5-api/actions/runs/4045702976/jobs/6958325313

Using GitHub Actions for releases checks out the release tag, which
bypasses this logic path.