Update packages for ccdb5-api and tests
The CCDB API requirements had been neglected for some time.
This PR brings dependencies up to date with the host cfgov framework, and restores the tox tests, which now pass.

Added files;
reqirements.txt file for local development and testing.

Note:
The API is limited to Python 3.8 because of restrictions on the current deployment environment.
That affects which version of Django we can use (can't move to Django 5).