Fix release wheel builds/uploads
The GitHub Action that builds wheels on releases has been broken for
some time due to Python 3.8's EOL, see for example
https://github.com/cfpb/ccdb5-api/actions/runs/17275449040

This commit updates the release building code to use a more modern
version of Python (3.13) - we don't need to use the runtime version of
Python (3.8) to build the wheel.

It also updates the action to run on pull requests to ensure that the
release package can be built (without actually doing the upload).