Remove outdated swagger-docs code and files
Our swagger docs rig has been a troublemaker for some time, because
the tech behind it has not been maintained and usage is questionable.

The only real options is to switch to a different documentation
framework, which we don't have resources to devote to at present.
This PR doesn't take down the original docs, which are still mostly relevant,
but it removes all the insecure code and files used to generate new versions.